Add middleware functionality to localstorage service
Middleware gives the option to manipulate data in a layer between redux and localstorage.
Can be used to filter out deeply nested properties not wanted in localstorage or to encrypt / decrypt data.

- applyMiddleware should be called to chain / sequence / compose multiple middlewares into one.
- getStorageForKeys will get data from storage and then pass them to middleware before passing to caller. This method should be used to rehydrate redux with data from localstorage on app startup.
- createPersistingStoreListener will run data through middleware before passing to setItem (persisting to localstorage). This method should be used to listen for redux changes and manipulate data before storing in localstorage.

The signature of a middleware looks like this:
const mw = (key, val) => val
The mw is called with the argument of the current storage key and the value for that key.
Expected return is the new/manipulated value for that key.

Two middlewares added:
- Add 'offline' bookmark (if it doesn't exist) when getting initial state.
- Set bookmarks.activeBookmark to 'offline' so app always starts in offline mode.