Cancel query on frame close / re-run click
To enable cancellation to be possible it was needed to refactor the structure of how we're storing requests and responses in Redux.
A `requestsDuck` is introduced and the frames in the existing `streamDuck` have a requestId attached to them.

A series of actions are now dispatched on every user command, and can be seen in the `commandsDuck.test.js` file.
Regarding the tests there, somehow Jest loses all actions dispatched within or after a Promise in a mergeMap. The code works but the tests don't. I've added a note in the test file about this.