Use server config browser.post_connect_cmd
On connection success, read config from server and look for `browser.post_connect_cmd`.
If found, execute the command as a system command (not adding it to command history).

I created a new test file because I needed to mock `bolt.transaction` for this new epic but needed the rest of bolt to be unmocked in the existing tests. Just creating a new file was the easiest way around this problem.