Set bolt host with URL param 'connectURL'
When Neo4j Browser is served from Neo4j Server the bolt url is auto discovered through a request to Neo4j Server `GET /`.
In the case of a externally hosted Neo4j Browser it can make connecting to Neo4j easier if a connect url can be passed along with the link to Neo4j Browser.

Example: `http://browser.neo4j.com/?connectURL=neo4jdb.xxyyzz.com:26000`.
If a hosted Neo4j Browser was available on `http://browser.neo4j.com` it would set the bolt url to `neo4jdb.xxyyzz.com:26000` and try to connect.

If no protocol is prepended the host, `bolt://` is assumed. If protocol is `bolt+routing://`, user settings `useBoltRouting` is set to `true` to enable routing when it's available.

A refactor of how to get the current url is in this as well.
To make the epics testable the action `APP_START` passes along the current url.