Remove standard in favor of eslint
Keep the standard config. Nothing will change style wise, we just use `eslint` directly rather than going through `standard`.

Move to `prettier-eslint` as well.

The point of this is to being able to configure the linting rules and extenstions in a simpler way.
With this commit we also bring in babel support for `stage-2` proposals so we don’t explicitly need to bring in `"transform-object-rest-spread"` and `”babel-plugin-transform-class-properties”` which is why I started this in the first place.