fix: Remove routing context when converts from neo4j to bolt protocol
Browser during the loggin is creating direct connection to the informed `connect url`, but it's not removing the query string from the url. Because of this, browser is getting a `Parameters are not supported with none route scheme..` error. The routing context params are not support in direct connection by design.

The solution for this issue is removing the routing context (query string) from the url when transalate from `neo4j` to `bolt`.