Added zoom to fit to space button for graph
Also added new icons for zoom in/out and fixed styling according to latest designs.

Changed old icon buttons to have the same background states as in the new designs, however doesn't change the color yet, since the old icons are smaller and thinner it makes them more difficult to read. Remove the legacy color when all new icons are in place.