Final cleanup of styles
Also, removed deps "@trivago/prettier-plugin-sort-imports" since it reorders also imports with side effects, like the order of the import of the style files. Reordered the style imports back to as it was in release 4.3.2.
When this is released (with an option to not reorder imports with side effects) https://github.com/trivago/prettier-plugin-sort-imports/pull/111 we can add the plugin again.