Fixed styles that were set in last release
when the css files were reordered hence the root font size was wrong, now when we have changed back the order the styles the rooot font size is different so the styles needs to be fixed. Set to px for now, until a complete change from px to rem is done.