Update settings for neo4j 5.0
Some settings renamed in 5.0:
- metrics.namespaces.enabled -> server.metrics.namespaces.enabled
- metrics.prefix -> server.metrics.prefix
- clients.allow_telemetry -> client.allow_telemetry

Typed the settings that are saved in overall state.

Set authEnabled to true until setting is read to be on the safe side.

Added support for duration that includes all durations that are supported for https://neo4j.com/docs/operations-manual/4.4/reference/configuration-settings/#config_browser.credential_timeout